+++ src/libaktualizr/primary/results.h
@@ -55,7 +55,24 @@ class UpdateCheck {
 /**
  * Result of an attempt to pause or resume a download.
  */
-using Pause = Uptane::Fetcher::PauseRet;
+enum class PauseStatus {
+  /* The system was successfully paused or resumed */
+  kSuccess = 0,
+  /* The system was already paused */
+  kAlreadyPaused,
+  /* The system was already running */
+  kAlreadyRunning,
+  /* General error */
+  kError,
+};
+
+class Pause {
+ public:
+  Pause() = default;
+  Pause(PauseStatus status_in) : status(status_in) {}
+
+  PauseStatus status{PauseStatus::kSuccess};
+};
 
 /**
  * Status of an update download.